puhs

+++ src/Hibernate/MessageDAO.java
@@ -11,10 +11,22 @@ public class MessageDAO {
 		em.getTransaction().commit();
 	}
 	
-	public static void Supprimer_Message(EntityManager em ,Message m) {
-		em.createQuery("delete from Message where id = :id")
-		  .setParameter("id", m.getId())
-		  .executeUpdate();
+	public static boolean Verify_title(EntityManager em, String titre, Utilisateur u) {
+		return true; 
+	}
+	
+	public static void Supprimer_Message(EntityManager em ,Message m,Utilisateur u) {
+	boolean t=false; 
+		for (int i=0; i<u.getListe_Message().size();i++) {
+			if(u.getListe_Message().get(i)==m) {
+				t=true; 
+			}
+		}
+		if(t==true) {
+			em.createQuery("delete from Message where id = :id")
+			.setParameter("id", m.getId())
+			.executeUpdate();
+		}
 	}
 	
 	public static Message GetMessage_from_title(EntityManager em, String titre) {
